# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  ent.Antenatal:
    properties:
      ADDED_TIME:
        description: ADDEDTIME holds the value of the "ADDED_TIME" field.
        type: string
      edges:
        $ref: '#/definitions/ent.AntenatalEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the AntenatalQuery when eager-loading is set.
        type: object
      id:
        description: ID of the ent.
        type: integer
    type: object
  ent.AntenatalEdges:
    properties:
      getmom:
        $ref: '#/definitions/ent.Pregnant'
        description: GETMOM holds the value of the GETMOM edge.
        type: object
      getstatus:
        $ref: '#/definitions/ent.Babystatus'
        description: GETSTATUS holds the value of the GETSTATUS edge.
        type: object
      takecare:
        $ref: '#/definitions/ent.User'
        description: TAKECARE holds the value of the TAKECARE edge.
        type: object
    type: object
  ent.Babystatus:
    properties:
      STATUS_BABY_NAME:
        description: STATUSBABYNAME holds the value of the "STATUS_BABY_NAME" field.
        type: string
      edges:
        $ref: '#/definitions/ent.BabystatusEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the BabystatusQuery when eager-loading is set.
        type: object
      id:
        description: ID of the ent.
        type: integer
    type: object
  ent.BabystatusEdges:
    properties:
      setstatus:
        description: SETSTATUS holds the value of the SETSTATUS edge.
        items:
          $ref: '#/definitions/ent.Antenatal'
        type: array
    type: object
  ent.Pregnant:
    properties:
      PREGNANT_AGE:
        description: PREGNANTAGE holds the value of the "PREGNANT_AGE" field.
        type: integer
      PREGNANT_NAME:
        description: PREGNANTNAME holds the value of the "PREGNANT_NAME" field.
        type: string
      edges:
        $ref: '#/definitions/ent.PregnantEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the PregnantQuery when eager-loading is set.
        type: object
      id:
        description: ID of the ent.
        type: integer
    type: object
  ent.PregnantEdges:
    properties:
      setmom:
        description: SETMOM holds the value of the SETMOM edge.
        items:
          $ref: '#/definitions/ent.Antenatal'
        type: array
    type: object
  ent.User:
    properties:
      USER_EMAIL:
        description: USEREMAIL holds the value of the "USER_EMAIL" field.
        type: string
      USER_NAME:
        description: USERNAME holds the value of the "USER_NAME" field.
        type: string
      edges:
        $ref: '#/definitions/ent.UserEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the UserQuery when eager-loading is set.
        type: object
      id:
        description: ID of the ent.
        type: integer
    type: object
  ent.UserEdges:
    properties:
      care:
        description: CARE holds the value of the CARE edge.
        items:
          $ref: '#/definitions/ent.Antenatal'
        type: array
    type: object
  gin.H:
    additionalProperties: true
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server for SUT SE 2563
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: SUT SA Example API
  version: "1.0"
paths:
  /antenatals:
    get:
      description: list antenatal entities
      operationId: list-antenatal
      parameters:
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/ent.Antenatal'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      summary: List antenatal entities
    post:
      consumes:
      - application/json
      description: Create antenatal
      operationId: create-antenatal
      parameters:
      - description: Antenatal entity
        in: body
        name: antenatal
        required: true
        schema:
          $ref: '#/definitions/ent.Antenatal'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ent.Antenatal'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Create antenatal
  /antenatals/{id}:
    delete:
      description: get antenatal by ID
      operationId: delete-antenatal
      parameters:
      - description: Antenatal ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Delete a antenatal entity by ID
    get:
      description: get antenatal by ID
      operationId: get-antenatal
      parameters:
      - description: Antenatal ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ent.Antenatal'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Get a antenatal entity by ID
    put:
      consumes:
      - application/json
      description: update antenatal by ID
      operationId: update-antenatal
      parameters:
      - description: Antenatal ID
        in: path
        name: id
        required: true
        type: integer
      - description: Antenatal entity
        in: body
        name: antenatal
        required: true
        schema:
          $ref: '#/definitions/ent.Antenatal'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ent.Antenatal'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Update a antenatal entity by ID
  /babystatuss:
    get:
      description: list babystatus entities
      operationId: list-babystatus
      parameters:
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/ent.Babystatus'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      summary: List babystatus entities
    post:
      consumes:
      - application/json
      description: Create babystatus
      operationId: create-babystatus
      parameters:
      - description: Babystatus entity
        in: body
        name: babystatus
        required: true
        schema:
          $ref: '#/definitions/ent.Babystatus'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ent.Babystatus'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Create babystatus
  /babystatuss/{id}:
    delete:
      description: get babystatus by ID
      operationId: delete-babystatus
      parameters:
      - description: Babystatus ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Delete a babystatus entity by ID
    get:
      description: get babystatus by ID
      operationId: get-babystatus
      parameters:
      - description: Babystatus ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ent.Babystatus'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Get a babystatus entity by ID
    put:
      consumes:
      - application/json
      description: update babystatus by ID
      operationId: update-babystatus
      parameters:
      - description: Babystatus ID
        in: path
        name: id
        required: true
        type: integer
      - description: Babystatus entity
        in: body
        name: babystatus
        required: true
        schema:
          $ref: '#/definitions/ent.Babystatus'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ent.Babystatus'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Update a babystatus entity by ID
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
  BasicAuth:
    type: basic
  OAuth2AccessCode:
    authorizationUrl: https://example.com/oauth/authorize
    flow: accessCode
    scopes:
      admin: ' Grants read and write access to administrative information'
    tokenUrl: https://example.com/oauth/token
    type: oauth2
  OAuth2Application:
    flow: application
    scopes:
      admin: ' Grants read and write access to administrative information'
      write: ' Grants write access'
    tokenUrl: https://example.com/oauth/token
    type: oauth2
  OAuth2Implicit:
    authorizationUrl: https://example.com/oauth/authorize
    flow: implicit
    scopes:
      admin: ' Grants read and write access to administrative information'
      write: ' Grants write access'
    type: oauth2
  OAuth2Password:
    flow: password
    scopes:
      admin: ' Grants read and write access to administrative information'
      read: ' Grants read access'
      write: ' Grants write access'
    tokenUrl: https://example.com/oauth/token
    type: oauth2
swagger: "2.0"